requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
scikit-learn
streamlit
nltk
